+++ src/client/GraphQLClient.ts
@@ -1,3 +1,4 @@
+// keep granular imports here fore browser build
 import type { ExecutionResult } from "graphql";
 import { toError } from "../util/toError";
 import { GraphQLRequestError } from "./GraphQLRequestError";
@@ -5,12 +6,11 @@ import { GraphQLRequestError } from "./GraphQLRequestError";
 export interface GraphQLClientOptions<TTypings extends GraphQLClientTypings> {
   url: string;
   fetch?: typeof fetch;
-  init?: BasicRequestInit;
-  operations?: TTypings["OperationNameToDocument"];
+  init?: GraphQLRequestInit;
+  typings?: TTypings;
 }
 
 export interface GraphQLClientTypings {
-  OperationName: string | never;
   QueryName: string | never;
   MutationName: string | never;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -20,13 +20,13 @@ export interface GraphQLClientTypings {
   OperationNameToDocument: Record<string, string>;
 }
 
-export interface GraphQLRequestOptions extends BasicRequestInit {
+export interface GraphQLRequestOptions extends GraphQLRequestInit {
   query: string;
   variables?: Record<string, unknown>;
   operationName?: string;
 }
 
-export type BasicRequestInit = Omit<
+export type GraphQLRequestInit = Omit<
   RequestInit,
   "headers" | "body" | "method"
 > & {
@@ -38,24 +38,28 @@ export class GraphQLClient<
 > {
   protected url: string;
   protected customFetch?: typeof fetch;
-  protected init?: BasicRequestInit;
+  protected init: GraphQLRequestInit;
   protected operations: TTypings["OperationNameToDocument"];
 
   constructor(options: GraphQLClientOptions<TTypings>) {
     this.url = options.url;
     this.customFetch = options.fetch;
-    this.init = options.init;
-    this.operations = options.operations ?? {};
+    this.init = { ...options.init, headers: { ...options.init?.headers } };
+    this.operations = options.typings?.OperationNameToDocument ?? {};
   }
 
-  async post<
+  setHeaders(headers: Record<string, string>) {
+    this.init.headers = { ...this.init.headers, ...headers };
+  }
+
+  async postNamed<
     TOperationName extends TTypings["QueryName"] | TTypings["MutationName"]
   >(
     operationName: TOperationName,
     variables: TTypings["OperationNameToVariables"][TOperationName],
-    init?: BasicRequestInit
+    init?: GraphQLRequestInit
   ) {
-    return this.postRaw<TTypings["OperationNameToData"][TOperationName]>({
+    return this.post<TTypings["OperationNameToData"][TOperationName]>({
       ...init,
       query: this.operations[operationName],
       operationName,
@@ -63,16 +67,16 @@ export class GraphQLClient<
     });
   }
 
-  async postRaw<TData>({
+  async post<TData>({
     query,
     variables,
     operationName,
     headers,
     ...init
   }: GraphQLRequestOptions) {
-    const f = this.customFetch ?? fetch;
+    const fetch_ = this.customFetch ?? fetch;
 
-    const response = await f(this.url, {
+    const response = await fetch_(this.url, {
       ...this.init,
       ...init,
       method: "POST",