+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fall back to `unknown` for unmapped scalars
 - Add `OperationTypings` in client-side code generation for simplified client typing
 - Add `fetch`-based, type-safe `GraphQLClient` for queries and mutations
+- Add type-safe `subscribeNamed` method to `GraphQLWebSocketClient`
 
 ## v1.0.0